backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.11.0
requests==2.31.0
scipy>=1.13.0
networkx>=3.0
numpy>=1.26.0
PyYAML>=6.0
datasets>=2.14.0
